Fixed immediate source and true source being inverted
Fixes crashes and incompatibilities with mods, plugins, and Minecraft itself.

Javadocs:

getImmediateSource: Retrieves the immediate causer of the damage, e.g. the arrow entity, not its shooter.
getTrueSource: Retrieves the true causer of the damage, e.g. the player who fired an arrow, the shulker who fired the bullet, etc.